requirements.txt
streamlit==1.22.0
pandas==1.5.0
numpy==1.23.4
scikit-learn==1.1.3
lxml==4.9.1
psycopg2-binary==2.9.5
sqlalchemy==1.4.44
joblib==1.2.0
python-dotenv==0.21.0
plotly==5.11.0